requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
transformers
torch
aiofiles